Kenzie/Sub-MAF-07: updated architecture
