+++ app/src/main/java/com.terracode.blueharvest/AccessibilitySettingsActivity.kt
@@ -17,10 +17,10 @@ import com.terracode.blueharvest.listeners.ColorSchemeListener
 import com.terracode.blueharvest.listeners.LanguageSelectionListener
 import com.terracode.blueharvest.listeners.TextSizeChangeListener
 import com.terracode.blueharvest.listeners.UnitToggleListener
-import com.terracode.blueharvest.utils.LocaleHelper
-import com.terracode.blueharvest.utils.SetTextSize
+import com.terracode.blueharvest.managers.LocaleManager
+import com.terracode.blueharvest.managers.TextSizeManager
 import com.terracode.blueharvest.utils.TextConstants
-import com.terracode.blueharvest.utils.ThemeHelper
+import com.terracode.blueharvest.managers.ThemeManager
 
 //Activity class for the accessibility setting page.
 class AccessibilitySettingsActivity : AppCompatActivity() {
@@ -40,15 +40,15 @@ class AccessibilitySettingsActivity : AppCompatActivity() {
         sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this)
 
         //Set setting values before setting the content view
-        val currentTheme = ThemeHelper.getCurrentTheme(this)
-        ThemeHelper.setColorOverlayTheme(this, currentTheme)
+        val currentTheme = ThemeManager.getCurrentTheme(this)
+        ThemeManager.setColorOverlayTheme(this, currentTheme)
 
         //Set the view
         setContentView(R.layout.activity_accessibility_settings)
 
         //Set the text size for the view onCreate
         val rootView = findViewById<View>(android.R.id.content).rootView
-        SetTextSize.setTextSizeView(this, rootView)
+        TextSizeManager.setTextSizeView(this, rootView)
 
         //Set the toolbar
         val toolbar: Toolbar = findViewById(R.id.settingsToolbar)
@@ -84,8 +84,8 @@ class AccessibilitySettingsActivity : AppCompatActivity() {
 
         //-----Logic for Language Spinner in Activity-----//
         val currentLanguagePosition = sharedPreferences.getInt("selectedLanguagePosition", 0)
-        val languagePosition = LocaleHelper.getLanguageCode(currentLanguagePosition)
-        LocaleHelper.setLocale(this, languagePosition)
+        val languagePosition = LocaleManager.getLanguageCode(currentLanguagePosition)
+        LocaleManager.setLocale(this, languagePosition)
 
         val languages = resources.getStringArray(R.array.languageArray)
         val languageAdapter = ArrayAdapter(
