+++ app/src/main/java/com.terracode.blueharvest/HomeActivity.kt
@@ -10,11 +10,11 @@ import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.preference.PreferenceManager
-import com.terracode.blueharvest.utils.LocaleHelper
+import com.terracode.blueharvest.managers.LocaleManager
 import com.terracode.blueharvest.utils.ReadJSONObject
 import com.terracode.blueharvest.utils.UnitConverter
-import com.terracode.blueharvest.utils.SetTextSize
-import com.terracode.blueharvest.utils.ThemeHelper
+import com.terracode.blueharvest.managers.TextSizeManager
+import com.terracode.blueharvest.managers.ThemeManager
 
 class HomeActivity : AppCompatActivity() {
 
@@ -39,12 +39,12 @@ class HomeActivity : AppCompatActivity() {
         sharedPreferences = PreferenceManager.getDefaultSharedPreferences(this)
 
         //Set setting values before setting the content view
-        val currentTheme = ThemeHelper.getCurrentTheme(this)
-        ThemeHelper.setColorOverlayTheme(this, currentTheme)
+        val currentTheme = ThemeManager.getCurrentTheme(this)
+        ThemeManager.setColorOverlayTheme(this, currentTheme)
 
         val currentLanguagePosition = sharedPreferences.getInt("selectedLanguagePosition", 0)
-        val languagePosition = LocaleHelper.getLanguageCode(currentLanguagePosition)
-        LocaleHelper.setLocale(this, languagePosition)
+        val languagePosition = LocaleManager.getLanguageCode(currentLanguagePosition)
+        LocaleManager.setLocale(this, languagePosition)
 
         //Set the view
         setContentView(R.layout.activity_home)
@@ -55,7 +55,7 @@ class HomeActivity : AppCompatActivity() {
 
         //Set the text size
         val rootView = findViewById<View>(android.R.id.content).rootView
-        SetTextSize.setTextSizeView(this, rootView)
+        TextSizeManager.setTextSizeView(this, rootView)
 
         //Set the declared TextView values equal to the IDs in the HomeActivity XML file.
         optimalRakeHeightTextView = findViewById(R.id.optimalRakeHeightValue)
